Calling `df.convert { }.with { }` from compiler plugin won't cause static initialization of TimeZone class
